requirements.txt
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.18.0
requests>=2.31.0
python-dateutil>=2.8.2
sqlalchemy>=2.0.0
rapidfuzz>=3.0.0
scipy>=1.10.0
openai>=1.0.0
beautifulsoup4>=4.12.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
